+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/src/Audacia.CodeAnalysis.Analyzers/Common/DiagnosticId.cs
@@ -21,5 +21,6 @@
         public const string UseTypedResultsInsteadOfIActionResult = "ACL1016";
         public const string AvoidBooleanParameters = "ACL1017";
         public const string SuppressionRequiresJustification = "ACL1018";
+        public const string AssertionScopeForMultipleAssertions = "ACL1019";
     }
 }